create people team icon component
