+++ src/components/PeopleListItem.vue
@@ -1,24 +1,11 @@
 <script setup lang="ts">
 import { BREAKPOINTS } from '@/styles/variables'
-
 import type { IUser } from '@/mockdata/users'
 import { useWindowResize } from '@/composables/useWindowResize'
-import { computed, ComputedRef } from 'vue'
-import { storeToRefs } from 'pinia'
-import { useTeams } from '@/stores/teams'
-import { userAddressMock } from '@/mockdata/usersAddress'
+import { computed } from 'vue'
+import PeopleListItemTeamIcon from './PeopleListItemTeamIcon.vue'
 
-const { teams } = storeToRefs(useTeams())
 const { width } = useWindowResize()
-
-const locale: 'en' | 'nl' = 'en'
-
-// const POSITION = {
-//   first: 'fist',
-//   last: 'last',
-//   single: 'single'
-// } as const
-// type TPosition = (typeof POSITION)[keyof typeof POSITION]
 
 const props = defineProps<{
   user: IUser
@@ -33,64 +20,8 @@ const dynamicClasses = computed((): string[] => {
   return classes
 })
 
-interface ITeamIconData {
-  label: string
-  backgroundColor: string
-  title: string
-}
-
-const getTeamIconData = (teamId: number): ITeamIconData | null => {
-  const hasTeam = props.user.teamIds.some((id) => id === teamId)
-  if (hasTeam) {
-    const team = teams.value[teamId]
-    return {
-      label: team.abbreviation,
-      backgroundColor: team.color,
-      title: team.name[locale]
-    }
-  } else return null
-}
-
-const iconData = computed((): ITeamIconData[] => {
-  const placeholderIconData: ITeamIconData = {
-    label: '',
-    backgroundColor: 'white',
-    title: ''
-  }
-  const data = [placeholderIconData, placeholderIconData, placeholderIconData]
-  const team1 = getTeamIconData(1)
-  const team2 = getTeamIconData(2)
-  const team3 = getTeamIconData(3)
-  if (team1) data[0] = team1
-  if (team2) data[1] = team2
-  if (team3) data[2] = team3
-  return data
-})
-
-// const teamIconData1 = computed(():ITeamIconData | null =>
-//   getTeamIconData(1)
-// )
-// const teamIconData2: ComputedRef<ITeamIconData | > = computed(() =>
-//   getTemIconData(2)
-// )
-// const teamIconData3: ComputedRef<ITeamIconData> = computed(() =>
-//   getTemIconData(3)
-// )
-
-// const circleOneLabel = computed(() => {
-//   const hasTeam1 = props.user.teamIds.some((id) => id === 1)
-//   if (hasTeam1) {
-//     return teams.value[1].name[locale]
-//   }
-//   return ''
-// })
-
 const onClick = () =>
   console.warn('Currentelly there is no functionalitty attached to this button')
-
-//TODO: think about a util function to get the team from the store.
-// Apply dynamic style color with the color from the team
-// Lable should come from the team
 </script>
 
 <template>
@@ -113,17 +44,8 @@ const onClick = () =>
         ><span>{{ user.phoneNumber }}</span>
       </div>
     </div>
-    <span class="team-icon"
-      ><!--I'm assuming the team will be always 3-->
-      <div
-        v-for="(item, index) in iconData"
-        :key="index"
-        :title="item.title"
-        class="circle"
-        :style="{ backgroundColor: item.backgroundColor }"
-      >
-        {{ item.label }}
-      </div>
+    <span class="team-icon">
+      <PeopleListItemTeamIcon :user="user" />
     </span>
     <span class="remove-team-member" @click="onClick">
       <carbon:trash-can class="w-5 h-5" />
@@ -165,23 +87,6 @@ const onClick = () =>
     display: flex;
     flex-direction: column;
 
-    @media screen and (max-width: $desktop) {
-      .divisor {
-        margin: 0 0.5rem;
-      }
-    }
-
-    @media screen and (min-width: $desktop) {
-      flex: 10;
-      flex-direction: row !important;
-      .divisor {
-        display: none;
-      }
-      &__item {
-        margin: 0 2rem;
-      }
-    }
-
     .username {
       font-weight: 600;
     }
@@ -193,7 +98,24 @@ const onClick = () =>
         margin-right: $margin-xs;
       }
     }
+
+    @media screen and (max-width: $desktop) {
+      .divisor {
+        margin: 0 0.5rem;
+      }
+    }
+    @media screen and (min-width: $desktop) {
+      flex: 10;
+      flex-direction: row !important;
+      .divisor {
+        display: none;
+      }
+      &__item {
+        margin: 0 2rem;
+      }
+    }
   }
+
   .team-icon {
     flex: 2;
     cursor: default;
@@ -203,19 +125,8 @@ const onClick = () =>
     @media screen and (min-width: $desktop) {
       flex: 0.5;
     }
-
-    .circle {
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      width: 24px;
-      height: 24px;
-      border-radius: 50%;
-      background-color: white;
-      border: 1px solid black;
-      margin-right: -5px;
-    }
   }
+
   .remove-team-member {
     flex: 1.4;
     display: flex;
@@ -228,4 +139,3 @@ const onClick = () =>
   }
 }
 </style>
-../views/mockdata/mockdata
